+++ mpc_energy/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.8.0
 * Added ability to setup multiple optional loads through Web UI
+* Streamlit crashing fix???
 
 ## 0.7.0
 * Added net grid interaction kWh sensor to aid automations.